improvement!: much better recursive DSL ergonomics
To leverage this, remove all nested entities from your recursive DSLs
that you've only added to be copies of the steps available in the outer
context.

The idea now behind recursive DSLs is that we don't "unimport"
recursive DSLs ever (because they could be used at any point),
and we also don't unimport any outer context when opening
a recursive DSL. This may need to be more involved in the future,
if some DSLs have conflicts, but this should set the stage for that.
For example, there may need to be an option to say
`only_entities: [:foo, :bar]` in a recursive DSL. This would require tracking
all entity names to modules at each given step of the DSL, but it can be
done.

As a bonus this should vastly improve large recursive DSL compile times as
far fewer modules must be compiled in those cases.

improvement: better async compiler task handling

the way we were doing it before we were not using `Task.await` which meant
certain crashes would just hang the compiler (fun)